fixed a number of problems
When I  made the parent Player and GameGenerator classes, I was messy and had some members shadow the other members in the base class.  Which wouldn't be a problem but I put Players in the blocking queue of the parent class then went looking for them in the (overriding) blocking queue of the subclass.  Fixed that.

Then there was this thing.  I put ArenaPlayers in the blocking queue that accepted Players.  So when I popped them off, they were Players.  The magic was to cast them back to ArenaPlayers.